deb: Bringup on cm-10.2
Proposed bringup for deb (Nexus 7 2013 LTE) on cm-10.2.
Contains improvements to flo as of this commit.
I would like to create a common repo or somehow merge certain
pieces together (init improvements for example) but I think
we should have an updated "official" deb repo first to start
that work.

Change-Id: I77ee8f419def673014a086d8338e5c947373b556