Update patch set 1
Patch Set 1:

No, I haven't.  I don't have access to the required hardware for it (SlimPort cable or Miracast dongle, etc).  If it works on flo, I'd imagine it works here.

But more importantly, is this not what I said in my comment: "... would like to get this to a greater audience."  ie. to people who can download this changeset and do have access to that hardware?

I thought code review and testing was the point of Gerrit Code Review.

Patch-set: 1