Update patch set 1
Patch Set 1:

I think you are jumping to some conclusions here. I am not attacking you or your device. These are simple questions that I need to ask. They only require simple answers.

I know exactly what the point of code review is. My above -1 is to hold this project in place until I have done a review on it. It is not a "it won't happen". However, taking such an assuming position does not make me want to continue a review and remove the -1 when I am done. (As I do during reviews.) Rather, it makes me want to change it to a -2 to spite your position.

Now, lets start over here. I put a -1 on the project to hold it in place while a review happens. Shall we continue the review?

Patch-set: 1