Update patch set 3
Patch Set 3: Verified+1 Code-Review+2

Voice calling on deb when network support is available will not be a requirement for this device to be merged. This is a tablet after all. Not a phone. As soon as the device is merged, that should be marked as a feature request on JIRA.

Everything is looking good here. Scott, email me to get some more contact going. jeagoss_(a)T_cyanogenmod_#dot+_org

Patch-set: 3
Label: Code-Review=+2