Remove /data/app/sensor_ctl_socket on boot.
Unix domain sockets need to be unlinked and re-created on each
reboot regardless, and removing the old socket left by a prior boot
ensures that we do not have a dead socket file in the wrong security
context, thereby yielding denials and preventing proper removal
and re-creation by the sensors service.

This same patch was applied to flo in commit
https://android-review.googlesource.com/79110 but we forgot to
apply it to deb.

Bug: 12728711
Change-Id: Ibe7b44d7be4659806340b71a8ffb6fe977333cb9