deb: fixes for RIL
- Allow Telephony to access /proc/cmdline, which it's been requesting
since 4.3

- Allow SIM hotswap. It is accessible and works just fine

- Increase radio scan timeout. Most devices set this to infinite, and
there's no reason to restrict to only 9 seconds.

- Explicitly ignore CDMA subscription on SIM card. This fixes RIL not
being recognized on some Verizon SIMs because the RIL stalls trying to
do CDMA registration. This isn't a perfect solution, as you still need
to cycle airplane mode or eject and reinsert the SIM to get it to work,
but it's better than no RIL. Requires
http://review.cyanogenmod.org/#/c/85250/

- In the same vein, ensure we are always in GSM mode as the RIL can get
confused and try to set the wrong network mode on Verizon. Requires
http://review.cyanogenmod.org/#/c/86646/
http://review.cyanogenmod.org/#/c/86586/ and dependencies thereof.

Change-Id: I75085f5fd56327a27203252801755109fe57117b